refactor: move ValueWithLiveness to utils.py, reset shared state on unmount
- Move ValueWithLiveness and _value_or_call (renamed to value_or_call) from RenderContext.py to utils.py
- Reset shared store value to None when all subscribers unmount to release resources
- Re-resolve initial value (re-calling callable if provided) on first new subscribe after reset
- Update test to expect initializer re-invocation after full unmount/reset
