+++ plugins/ui/src/deephaven/ui/_internal/utils.py
@@ -5,6 +5,7 @@ from typing import (
     Dict,
     List,
     Mapping,
+    Generic,
     Set,
     Tuple,
     cast,
@@ -19,8 +20,10 @@ from deephaven.dtypes import (
 )
 from inspect import signature
 import sys
+from dataclasses import dataclass
 from functools import partial
 from itertools import zip_longest
+from deephaven.liveness_scope import LivenessScope
 from deephaven.time import (
     to_j_instant,
     to_j_zdt,
@@ -43,6 +46,36 @@ from ..types import (
 )
 
 T = TypeVar("T")
+
+
+@dataclass
+class ValueWithLiveness(Generic[T]):
+    """A value with an associated liveness scope, if any."""
+
+    value: T
+    liveness_scope: Union[LivenessScope, None]
+
+
+def value_or_call(
+    value: T | None | Callable[[], T | None],
+) -> ValueWithLiveness[T | None]:
+    """
+    Creates a wrapper around the value, or invokes a callable to hold the value and the liveness scope
+    creates while obtaining that value.
+
+    Args:
+        value: a value, or callable that will produce a value
+
+    Returns:
+        The resulting value, plus a liveness scope, if any.
+    """
+    if callable(value):
+        scope = LivenessScope()
+        with scope.open():
+            value = value()
+        return ValueWithLiveness(value=value, liveness_scope=scope)
+    return ValueWithLiveness(value=value, liveness_scope=None)
+
 
 _UNSAFE_PREFIX = "UNSAFE_"
 _ARIA_PREFIX = "aria_"
